Create xcodebuild package.
